+++ config/wallaby-mocha-ts.ts
@@ -23,7 +23,8 @@ module.exports = function (wallaby) {
       const mocha = wallaby.testFramework;
       mocha.timeout(30000);
       process.env.IN_WALLABY = true;
-      require('yoshi/config/test-setup');
+      require('./test-setup');
+
     },
     env: {
       type: 'node',