+++ config/wallaby-mocha.js
@@ -6,7 +6,7 @@ module.exports = function (wallaby) {
     const mocha = wallaby.testFramework;
     mocha.timeout(30000);
     process.env.IN_WALLABY = true;
-    require('yoshi/config/test-setup');
+    require('./test-setup');
   };
   return wallabyCommon;
 };