+++ config/webpack.config.client.js
@@ -10,7 +10,8 @@ const {BundleAnalyzerPlugin} = require('webpack-bundle-analyzer');
 const DuplicatePackageCheckerPlugin = require('duplicate-package-checker-webpack-plugin');
 const {isObject} = require('lodash');
 
-const defaultCommonsChunkConfig = {
+const defaultSplitChunksConfig = {
+  chunks: 'all',
   name: 'commons',
   minChunks: 2
 };
@@ -20,8 +21,8 @@ const config = ({debug, separateCss = projectConfig.separateCss(), analyze, disa
   const projectName = projectConfig.name();
   const cssModules = projectConfig.cssModules();
   const tpaStyle = projectConfig.tpaStyle();
-  const useCommonsChunk = projectConfig.commonsChunk();
-  const commonsChunkConfig = isObject(useCommonsChunk) ? useCommonsChunk : defaultCommonsChunkConfig;
+  const useSplitChunks = projectConfig.splitChunks();
+  const splitChunksConfig = isObject(useSplitChunks) ? useSplitChunks : defaultSplitChunksConfig;
 
   if (separateCss === 'prod') {
     if (inTeamCity() || isProduction()) {
@@ -34,6 +35,13 @@ const config = ({debug, separateCss = projectConfig.separateCss(), analyze, disa
   return mergeByConcat(webpackConfigCommon, {
     entry: getEntry(),
 
+    mode: debug ? 'development' : 'production',
+
+    optimization: {
+      minimize: !debug,
+      splitChunks: useSplitChunks ? splitChunksConfig : false,
+    },
+
     module: {
       rules: [
         ...require('../src/loaders/sass')(separateCss, cssModules, tpaStyle, projectName).client,
@@ -44,7 +52,6 @@ const config = ({debug, separateCss = projectConfig.separateCss(), analyze, disa
     plugins: [
       ...disableModuleConcat ? [] : [new webpack.optimize.ModuleConcatenationPlugin()],
       ...analyze ? [new BundleAnalyzerPlugin()] : [],
-      ...useCommonsChunk ? [new webpack.optimize.CommonsChunkPlugin(commonsChunkConfig)] : [],
 
       new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
 
@@ -64,15 +71,6 @@ const config = ({debug, separateCss = projectConfig.separateCss(), analyze, disa
       ...!separateCss ? [] : [
         new ExtractTextPlugin(debug ? '[name].css' : '[name].min.css'),
         new RtlCssPlugin(debug ? '[name].rtl.css' : '[name].rtl.min.css'),
-      ],
-
-      ...debug ? [] : [
-        new webpack.optimize.UglifyJsPlugin({
-          sourceMap: true,
-          compress: {
-            warnings: false,
-          },
-        })
       ],
     ],