+++ src/loaders/ng-annotate.js
@@ -1,0 +1,7 @@
+const {unprocessedModules} = require('../../config/project');
+
+module.exports = () => ({
+  test: /\.js?$/,
+  loader: 'ng-annotate-loader',
+  include: unprocessedModules()
+});