+++ src/plugins/haste-plugin-yoshi-logger/index.js
@@ -1,5 +1,6 @@
 const chalk = require('chalk');
 const {format, delta, generateRunTitle} = require('./utils');
+const get = require('lodash/get');
 
 module.exports = class LoggerPlugin {
   apply(runner) {
@@ -17,13 +18,16 @@ module.exports = class LoggerPlugin {
           const start = new Date();
 
           const title = generateRunTitle({run, task});
+          const showLogs = get(run, 'runnerOptions.log', true);
 
-          console.log(`[${format(start)}] ${chalk.black.bgGreen('Starting')} '${title}'...`);
+          if (showLogs) {
+            console.log(`[${format(start)}] ${chalk.black.bgGreen('Starting')} '${title}'...`);
 
-          run.hooks.success.tap('log success', () => {
-            const [end, time] = delta(start);
-            console.log(`[${format(end)}] ${chalk.black.bgCyan('Finished')} '${title}' after ${time} ms`);
-          });
+            run.hooks.success.tap('log success', () => {
+              const [end, time] = delta(start);
+              console.log(`[${format(end)}] ${chalk.black.bgCyan('Finished')} '${title}' after ${time} ms`);
+            });
+          }
 
           run.hooks.failure.tap('log failure', () => {
             const [end, time] = delta(start);