+++ test/babel/fixtures/with-imports/wrap-import-named-render/expected.js
@@ -1,0 +1,4 @@
+import { AppContainer as _AppContainer } from 'react-hot-loader';
+import { render } from 'react-dom';
+
+render(<_AppContainer><Component /></_AppContainer>, x);