+++ test/babel/fixtures/with-imports/wrap-import-wildcard-render/actual.js
@@ -1,0 +1,3 @@
+import * as reactDOM from 'react-dom';
+
+reactDOM.render(<Component />, x);