+++ test/babel/fixtures/with-imports/wrap-import-wildcard-render/expected.js
@@ -1,0 +1,4 @@
+import { AppContainer as _AppContainer } from 'react-hot-loader';
+import * as reactDOM from 'react-dom';
+
+reactDOM.render(<_AppContainer><Component /></_AppContainer>, x);