+++ test/babel/fixtures/with-requires/no-wrap-required-variable-call-render/actual.js
@@ -1,0 +1,3 @@
+const reactDOM = 'react-dom';
+
+reactDOM(<Component />, x);