+++ test/babel/fixtures/with-requires/wrap-import-destructured-render/expected.js
@@ -1,0 +1,4 @@
+import { AppContainer as _AppContainer } from 'react-hot-loader';
+const { render } = require('react-dom');
+
+render(<_AppContainer><Component /></_AppContainer>, x);