+++ test/babel/fixtures/with-requires/wrap-required-method-from-variable-render/expected.js
@@ -1,0 +1,4 @@
+import { AppContainer as _AppContainer } from 'react-hot-loader';
+const reactDOM = require('react-dom');
+
+reactDOM.render(<_AppContainer><Component /></_AppContainer>, x);